filter: split headers into hpp + cpp
Move the implementations of CategoryIndex, NumericIndex, Bucket, and
Filter from their respective headers into new translation units. The
headers now expose only types, declarations, and the tiny inline
accessors (sortable_from_float family, Bucket::get_value /
is_full / is_empty). Behavior is unchanged; this is a build-time
refactor.

Define NDD_FILTER_SOURCES once in the root CMakeLists.txt and pull it
into both NDD_CORE_SOURCES (for the main binary) and the
ndd_filter_test target so the implementations are linked in both
places.

Add #include <thread> to settings.hpp. It uses
std::thread::hardware_concurrency() but was relying on a transitive
include from the old filter.hpp; the trimmed filter.hpp no longer pulls
in <thread>, so the test build broke without this fix.

Verified: ndd_filter_test (42 pass, 7 skip, 0 fail) and
ndd_request_validation_test (6 pass, 0 fail) match the pre-split
results; ndd-avx2 builds clean.
