+++ src/utils/settings.hpp
@@ -5,6 +5,7 @@
 #include <sstream>
 #include <algorithm>
 #include <cstdint>
+#include <thread>
 
 constexpr uint64_t KB = (1024ULL);
 constexpr uint64_t MB = (1024ULL * KB);
